@@ -34,6 +34,9 @@
     "prepack": "./scripts/build"
   },
   "dependencies": {
+    "@types/pg": "^7.4.10",
+    "@types/graphql": "^0.8.2",
+    "@types/jsonwebtoken": "<7.2.1",
     "body-parser": "^1.15.2",
     "chalk": "1.1.3",
     "commander": "2.9.0",
@@ -57,10 +60,8 @@
     "@types/change-case": "2.3.1",
     "@types/commander": ">=2.3.30 <2.12.0",
     "@types/debug": "0.0.29",
-    "@types/graphql": "^0.8.2",
     "@types/jest": "^16.0.4",
-    "@types/jsonwebtoken": "<7.2.1",
-    "@types/lodash": ">=4.14.30 <4.14.57",
+    "@types/lodash": "^4.14.110",
     "@types/node": "^7.0.4",
     "codemirror-graphql": "^0.6.12",
     "connect": "^3.5.0",
@@ -75,8 +76,8 @@
     "source-map-support": "^0.4.6",
     "supertest": "^2.0.1",
     "ts-node": "^2.0.0",
-    "tslint": "^4.2.0",
-    "typescript": "2.1.5"
+    "tslint": "^5.10.0",
+    "typescript": "^2.9.2"
   },
   "jest": {
     "transform": {
