@@ -11,10 +11,13 @@ import { Pool } from 'pg'
 import cluster = require('cluster')
 import { makePluginHook, PostGraphilePlugin } from './pluginHook'
 import debugFactory = require('debug')
+import { Postgraphile } from '../interfaces'
+import mixed = Postgraphile.mixed
 
 const debugCli = debugFactory('postgraphile:cli')
 
-// tslint:disable no-console
+// tslint:disable no-console no-shadowed-variable
+// no-shadowed-variable is erroroneously reporting
 
 let config = {}
 try {
