@@ -6,6 +6,6 @@ declare module 'pg-sql2' {
 
   export function compile(query: SQLQuery): QueryConfig
   export function query(strings: TemplateStringsArray, ...values: Array<SQLNode | SQLQuery>): SQLQuery
-  export function value(val: mixed): SQLNode
+  export function value(val: any): SQLNode
   export function join(items: Array<SQLNode | SQLQuery>, separator?: string): SQLQuery
 }
